:recycle: :fire: Remove create() and associated types and methods from register second passphrase
