+++ packages/lisk-transactions/src/transactions/1_register_second_passphrase.ts
@@ -12,7 +12,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import { getKeys, hexToBuffer } from '@liskhq/lisk-cryptography';
+import { hexToBuffer } from '@liskhq/lisk-cryptography';
 import * as BigNum from 'browserify-bignum';
 import { SIGNATURE_FEE } from '../constants';
 import { TransactionError, TransactionMultiError } from '../errors';
@@ -22,7 +22,7 @@ import {
 	Status,
 	TransactionJSON,
 } from '../transaction_types';
-import { createBaseTransaction, validator } from '../utils';
+import { validator } from '../utils';
 import { BaseTransaction, TransactionResponse } from './base';
 
 const TRANSACTION_SIGNATURE_TYPE = 1;
@@ -37,12 +37,6 @@ export interface SignatureObject {
 
 export interface SecondSignatureAsset {
 	readonly signature: SignatureObject;
-}
-
-export interface SecondSignatureInput {
-	readonly passphrase: string;
-	readonly secondPassphrase: string;
-	readonly timeOffset?: number;
 }
 
 export const secondSignatureAssetTypeSchema = {
@@ -78,12 +72,6 @@ export const secondSignatureAssetFormatSchema = {
 	},
 };
 
-const validateInputs = ({ secondPassphrase }: SecondSignatureInput): void => {
-	if (typeof secondPassphrase !== 'string') {
-		throw new Error('Please provide a secondPassphrase. Expected string.');
-	}
-};
-
 export class SecondSignatureTransaction extends BaseTransaction {
 	public readonly asset: SecondSignatureAsset;
 	public constructor(tx: TransactionJSON) {
@@ -107,30 +95,6 @@ export class SecondSignatureTransaction extends BaseTransaction {
 		}
 		this.asset = tx.asset as SecondSignatureAsset;
 		this._fee = new BigNum(SIGNATURE_FEE);
-	}
-
-	public static create(input: SecondSignatureInput): object {
-		validateInputs(input);
-		const { passphrase, secondPassphrase } = input;
-		const { publicKey } = getKeys(secondPassphrase);
-
-		const transaction = {
-			...createBaseTransaction(input),
-			type: 1,
-			fee: SIGNATURE_FEE.toString(),
-			asset: { signature: { publicKey } },
-		};
-
-		if (!passphrase) {
-			return transaction;
-		}
-
-		const secondSignatureTransaction = new SecondSignatureTransaction(
-			transaction as TransactionJSON,
-		);
-		secondSignatureTransaction.sign(passphrase, secondPassphrase);
-
-		return secondSignatureTransaction.toJSON();
 	}
 
 	protected getAssetBytes(): Buffer {
