+++ packages/lisk-transactions/test/transactions/1_register_second_passphrase.ts
@@ -23,8 +23,6 @@ import {
 	validTransaction,
 } from '../../fixtures';
 import { Status, TransactionJSON } from '../../src/transaction_types';
-import * as utils from '../../src/utils';
-import { SIGNATURE_FEE } from '../../src/constants';
 import { hexToBuffer } from '@liskhq/lisk-cryptography';
 
 describe('Second signature registration transaction class', () => {
@@ -63,42 +61,6 @@ describe('Second signature registration transaction class', () => {
 			expect(
 				() => new SecondSignatureTransaction(invalidSecondSignatureTransaction),
 			).to.throw('Invalid field types');
-		});
-	});
-
-	describe('#create', () => {
-		const timeWithOffset = 38350076;
-		const passphrase = 'secret';
-		const secondPassphrase = 'second secret';
-
-		let result: object;
-
-		beforeEach(async () => {
-			sandbox.stub(utils, 'getTimeWithOffset').returns(timeWithOffset);
-		});
-
-		describe('when the transaction is created with passphrase and second passphrase', () => {
-			beforeEach(async () => {
-				result = SecondSignatureTransaction.create({
-					passphrase,
-					secondPassphrase,
-				});
-			});
-
-			it('should create second signature registration transaction', async () => {
-				expect(result).to.have.property('id');
-				expect(result).to.have.property('type', 1);
-				expect(result).to.have.property('amount', '0');
-				expect(result).to.have.property('fee', SIGNATURE_FEE.toString());
-				expect(result).to.have.property('senderId');
-				expect(result).to.have.property('senderPublicKey');
-				expect(result).to.have.property('recipientId', '');
-				expect(result).to.have.property('timestamp', timeWithOffset);
-				expect(result).to.have.property('signature').and.not.to.be.empty;
-				expect((result as any).asset.signature.publicKey).to.eql(
-					'0401c8ac9f29ded9e1e4d5b6b43051cb25b22f27c7b7b35092161e851946f82f',
-				);
-			});
 		});
 	});
